# swagger.yaml
...
swagger: "2.0"
info:
  description: "The inner API of Popcube aim to define every methods need to manage\
    \ communication organisation inside PopCube chat project"
  version: "alpha"
  title: "Popcube iner API"
  contact:
    name: "Titouan Freville"
    email: "titouanfreville@gmail.com"
host: "api.popcube"
basePath: "/alpha"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /emoji:
    get:
      tags:
      - "Emojis"
      summary: "Get emojis"
      description: "This will get all the emojis available in the organisation."
      operationId: "getAllEmoji"
      parameters: []
      responses:
        200:
          description: "emojiSlice Array of emojis"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Emoji"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
    post:
      tags:
      - "Emojis"
      summary: "New emoji"
      description: "This will create an emoji for organisation emojis library."
      operationId: "newEmoji1"
      parameters:
      - in: "body"
        name: "emoji"
        description: "Link of the emoji in server."
        required: false
        schema:
          $ref: "#/definitions/Emoji"
        x-go-name: "Emoji"
      responses:
        200:
          description: "emojiObjectSuccess list of emojis"
          schema:
            $ref: "#/definitions/Emoji"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /emoji/all:
    get:
      tags:
      - "Emojis"
      summary: "Get emojis"
      description: "This will get all the emojis available in the organisation."
      operationId: "getAllEmoji1"
      parameters: []
      responses:
        200:
          description: "emojiSlice Array of emojis"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Emoji"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /emoji/link/{emojiLink}:
    get:
      tags:
      - "Emojis"
      summary: "Get emoji from link"
      description: "This will return the emoji object corresponding to provided link"
      operationId: "getEmojiFromLink"
      parameters:
      - name: "emojiLink"
        in: "path"
        description: "Link of the emoji in server."
        required: true
        type: "string"
        x-go-name: "EmojiLink"
      responses:
        200:
          description: "emojiObjectSuccess list of emojis"
          schema:
            $ref: "#/definitions/Emoji"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /emoji/name/{emojiName}:
    get:
      tags:
      - "Emojis"
      summary: "Get emoji from name"
      description: "This will return the emoji object corresponding to provided name"
      operationId: "getEmojiFromName"
      parameters:
      - name: "emojiName"
        in: "path"
        description: "Link of the emoji in server."
        required: true
        type: "string"
        x-go-name: "EmojiName"
      responses:
        200:
          description: "emojiObjectSuccess list of emojis"
          schema:
            $ref: "#/definitions/Emoji"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /emoji/shortcut/{emojiShortcut}:
    get:
      tags:
      - "Emojis"
      summary: "Get emoji from shortcut"
      description: "This will return the emoji object corresponding to provided shortcut"
      operationId: "getEmojiFromShortcut"
      parameters:
      - name: "emojiShortcut"
        in: "path"
        description: "Link of the emoji in server."
        required: true
        type: "string"
        x-go-name: "EmojiShortcut"
      responses:
        200:
          description: "emojiObjectSuccess list of emojis"
          schema:
            $ref: "#/definitions/Emoji"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /emoji/{emojiID}:
    put:
      tags:
      - "Emojis"
      summary: "Update emoji"
      description: "This will return the new emoji object"
      operationId: "updateEmoji"
      parameters:
      - name: "emojiID"
        in: "path"
        description: "Link of the emoji in server."
        required: true
        type: "integer"
        format: "int64"
        x-go-name: "EmojiID"
      - in: "body"
        name: "emoji"
        description: "Link of the emoji in server."
        required: false
        schema:
          $ref: "#/definitions/Emoji"
        x-go-name: "Emoji"
      responses:
        200:
          description: "avatarObjectSuccess list of avatars"
          schema:
            $ref: "#/definitions/Avatar"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
    delete:
      tags:
      - "Emojis"
      summary: "Delete emoji"
      description: "This will return an object describing the deletion"
      operationId: "deleteEmoji"
      parameters:
      - name: "emojiID"
        in: "path"
        description: "Link of the emoji in server."
        required: true
        type: "integer"
        format: "int64"
        x-go-name: "EmojiID"
      responses:
        200:
          description: "deleteMessage"
          schema:
            $ref: "#/definitions/deleteMessage"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
...
definitions:
  Avatar:
    type: "object"
    required:
    - "link"
    - "name"
    properties:
      id:
        type: "integer"
        format: "uint64"
        description: "id of the avatar"
        minimum: 0.0
        x-go-name: "IDAvatar"
      link:
        type: "string"
        description: "Path into server"
        x-go-name: "Link"
      name:
        type: "string"
        description: "Avatar name"
        x-go-name: "Name"
    title: "Avatar object"
    description: "Avatar object store default avatar images you can use within an\
      \ association.\nRequired apply only for creation of the object. Id is required\
      \ only on update."
    x-go-package: "github.com/titouanfreville/popcubeapi/models"
...
  Emoji:
    type: "object"
    required:
    - "link"
    - "name"
    - "shortcut"
    properties:
      id:
        type: "integer"
        format: "uint64"
        description: "id of the emoji"
        minimum: 0.0
        x-go-name: "IDEmoji"
      link:
        type: "string"
        description: "path to emoji into server"
        x-go-name: "Link"
      name:
        type: "string"
        maxLength: 64
        x-go-name: "Name"
      shortcut:
        type: "string"
        maxLength: 20
        x-go-name: "Shortcut"
    title: "Emoji object."
    description: "Emoji object describe the emoji available in the organisation and\
      \ theire shortcuts.\nRequired apply only for creation of the object."
    x-go-package: "github.com/titouanfreville/popcubeapi/models"
...
  deleteMessage:
    type: "object"
    properties:
      Message:
        description: "in:body"
        $ref: "#/definitions/deleteMessageModel"
    title: "deleteMessage return object to confirm correct deletion of an item."
    x-go-package: "github.com/titouanfreville/popcubeapi/api"
  deleteMessageModel:
    type: "object"
    properties:
      message:
        type: "string"
        description: "More information about why is it or isn't it removed"
        x-go-name: "Message"
      removed_object:
        type: "object"
        description: "The object we where trying to remove"
        properties: {}
        x-go-name: "Object"
      status:
        type: "integer"
        format: "int64"
        description: "Status"
        x-go-name: "Status"
      success:
        type: "boolean"
        description: "Correctly removed ?"
        x-go-name: "Success"
    title: "deleteMessageModel is an object to confirm correct deletion of an item."
    x-go-package: "github.com/titouanfreville/popcubeapi/api"
...
  genericErrorModel:
    type: "object"
    properties:
      detailed_error:
        type: "string"
        description: "Detail of error"
        x-go-name: "DetailedError"
      id:
        type: "string"
        description: "ID for the error"
        x-go-name: "ID"
      is_oauth:
        type: "boolean"
        x-go-name: "IsOAuth"
      message:
        type: "string"
        description: "Error's message"
        x-go-name: "Message"
      request_id:
        type: "string"
        description: "Id of the request if exist"
        x-go-name: "RequestID"
      status_code:
        type: "integer"
        format: "int64"
        description: "EgenericErrorModelrror's status code"
        x-go-name: "StatusCode"
    title: "AppError Type used to structure error reporting for popcube chat project."
    x-go-name: "AppError"
    x-go-package: "github.com/titouanfreville/popcubeapi/utils"
...